+++ package.json
@@ -28,7 +28,6 @@
     "fast-xml-parser": "^4.0.2",
     "json-tree-view-vue3": "^0.1.16",
     "moment": "^2.29.1",
-    "node-kcp-x": "^1.0.1",
     "onfire.js": "^2.0.1",
     "protobufjs": "^6.11.2",
     "redis": "^4.0.4",
@@ -44,8 +43,7 @@
     "cross-env": "^7.0.3",
     "electron": "16.0.6",
     "electron-builder": "^22.14.5",
-    "electron-rebuild": "^3.2.7",
-    "ts-node": "^10.8.0",
+    "ts-node": "^10.4.0",
     "typescript": "^4.5.4",
     "vite": "^2.7.10",
     "vite-plugin-resolve": "^1.3.1",
